Introduce School auth screen and route snapshots
Replace client-side school/class fetches with server-resolved route
snapshots passed into the School and Class context providers. Add a
shared SchoolAuthScreen for unauthenticated school entry and new
SchoolContentState/SchoolContentLoading components for consistent
loading/empty UI.

Split materials list into a dedicated item component and types, add
forum search parsers, academic-year helpers, and various UI/permission
tweaks across class screens. Add backend validators and unit tests for
class route contracts and strengthen query/mutation validation.
