+++ apps/www/app/[locale]/(app)/school/(authenticated)/[slug]/(main)/classes/[id]/layout.tsx
@@ -1,14 +1,19 @@
+import { captureServerException } from "@repo/analytics/posthog/server";
+import { api } from "@repo/backend/convex/_generated/api";
 import type { Id } from "@repo/backend/convex/_generated/dataModel";
 import { ErrorBoundary } from "@repo/design-system/components/ui/error-boundary";
+import { fetchQuery } from "convex/nextjs";
 import { use } from "react";
 import { SchoolClassesForumPostSheet } from "@/components/school/classes/forum/post-sheet";
 import { SchoolClassesHeaderInfo } from "@/components/school/classes/info";
+import { SchoolClassesJoinForm } from "@/components/school/classes/join-form";
 import { SchoolClassesTabs } from "@/components/school/classes/tabs";
-import { SchoolClassesValidation } from "@/components/school/classes/validation";
 import { SchoolNotFound } from "@/components/school/not-found";
+import { getToken } from "@/lib/auth/server";
 import { ClassContextProvider } from "@/lib/context/use-class";
 import { ForumContextProvider } from "@/lib/context/use-forum";
 
+/** Bind the resolved class route snapshot to the class subtree. */
 export default function Layout(
   props: LayoutProps<"/[locale]/school/[slug]/classes/[id]">
 ) {
@@ -17,10 +22,45 @@ export default function Layout(
 
   const classId = id as Id<"schoolClasses">;
 
-  return (
-    <ErrorBoundary fallback={<SchoolNotFound />}>
-      <SchoolClassesValidation classId={classId}>
-        <ClassContextProvider classId={classId}>
+  return <ClassRouteBoundary classId={classId}>{children}</ClassRouteBoundary>;
+}
+
+/**
+ * Resolve the class route snapshot on the server so the client subtree only
+ * consumes stable class state.
+ */
+async function ClassRouteBoundary({
+  children,
+  classId,
+}: {
+  children: React.ReactNode;
+  classId: Id<"schoolClasses">;
+}) {
+  const token = await getToken();
+
+  if (!token) {
+    return <SchoolNotFound />;
+  }
+
+  try {
+    const value = await fetchQuery(
+      api.classes.queries.getClassRoute,
+      { classId },
+      { token }
+    );
+
+    if (value.kind === "joinRequired") {
+      return (
+        <SchoolClassesJoinForm
+          classId={value.class._id}
+          visibility={value.class.visibility}
+        />
+      );
+    }
+
+    return (
+      <ErrorBoundary fallback={<SchoolNotFound />}>
+        <ClassContextProvider value={value}>
           <SchoolClassesHeaderInfo />
           <SchoolClassesTabs />
 
@@ -29,7 +69,14 @@ export default function Layout(
             {children}
           </ForumContextProvider>
         </ClassContextProvider>
-      </SchoolClassesValidation>
-    </ErrorBoundary>
-  );
+      </ErrorBoundary>
+    );
+  } catch (error) {
+    await captureServerException(error, undefined, {
+      classId,
+      source: "school-class-route-boundary",
+    });
+
+    return <SchoolNotFound />;
+  }
 }
