+++ apps/www/app/[locale]/(app)/school/page.tsx
@@ -1,20 +1,22 @@
-import { Particles } from "@repo/design-system/components/ui/particles";
-
 import { Suspense } from "react";
 import { School } from "@/components/school";
+import { SchoolAuthScreen } from "@/components/school/auth-screen";
 import { SchoolLoader } from "@/components/school/loader";
-import { ComingSoon } from "@/components/shared/coming-soon";
+import { getLocaleOrThrow } from "@/lib/i18n/params";
 
-export default function Page() {
+/**
+ * Resolve the school entry experience.
+ *
+ * Authenticated users are redirected into their school flow by `School`, while
+ * unauthenticated users see the shared Nakafa School sign-in screen.
+ */
+export default async function Page(props: PageProps<"/[locale]/school">) {
+  const locale = getLocaleOrThrow((await props.params).locale);
+
   return (
     <Suspense fallback={<SchoolLoader />}>
-      <School>
-        <div className="relative flex h-svh items-center justify-center">
-          <Particles className="pointer-events-none absolute inset-0 opacity-80" />
-          <div className="mx-auto w-full max-w-xl px-6">
-            <ComingSoon />
-          </div>
-        </div>
+      <School locale={locale}>
+        <SchoolAuthScreen />
       </School>
     </Suspense>
   );
