+++ apps/www/components/school/classes/forum/search-params.ts
@@ -1,0 +1,6 @@
+import { parseAsString } from "nuqs/server";
+
+/** Query-string parsers used by class forum routes and overlays. */
+export const forumSearchParsers = {
+  forum: parseAsString,
+};
