+++ packages/backend/convex/classes/queries.ts
@@ -1,12 +1,17 @@
 import { query } from "@repo/backend/convex/_generated/server";
 import { SCHOOL_CLASS_INVITE_CODE_ROLES } from "@repo/backend/convex/classes/constants";
 import {
-  classInfoValidator,
   paginatedClassesValidator,
   paginatedPeopleValidator,
   schoolClassVisibilityValidator,
 } from "@repo/backend/convex/classes/schema";
-import { loadClass } from "@repo/backend/convex/classes/utils";
+import { loadActiveClass, loadClass } from "@repo/backend/convex/classes/utils";
+import {
+  classAccessResultValidator,
+  classInfoValidator,
+  classMembershipResultValidator,
+  classRouteResultValidator,
+} from "@repo/backend/convex/classes/validators";
 import { requireAuth } from "@repo/backend/convex/lib/helpers/auth";
 import {
   checkClassAccess,
@@ -20,7 +25,6 @@ import { getUserMap } from "@repo/backend/convex/lib/helpers/user";
 import { vv } from "@repo/backend/convex/lib/validators/vv";
 import { paginationOptsValidator } from "convex/server";
 import { ConvexError, v } from "convex/values";
-import { nullable } from "convex-helpers/validators";
 
 const MAX_CLASS_MEMBER_SEARCH_RESULTS = 500;
 
@@ -146,7 +150,7 @@ export const verifyClassMembership = query({
   args: {
     classId: vv.id("schoolClasses"),
   },
-  returns: v.object({ allow: v.boolean() }),
+  returns: classAccessResultValidator,
   handler: async (ctx, args) => {
     const user = await requireAuth(ctx);
 
@@ -171,11 +175,7 @@ export const getClass = query({
   args: {
     classId: vv.id("schoolClasses"),
   },
-  returns: v.object({
-    class: vv.doc("schoolClasses"),
-    classMembership: nullable(vv.doc("schoolClassMembers")),
-    schoolMembership: vv.doc("schoolMembers"),
-  }),
+  returns: classMembershipResultValidator,
   handler: async (ctx, args) => {
     const user = await requireAuth(ctx);
 
@@ -190,6 +190,59 @@ export const getClass = query({
     return {
       class: classData,
       classMembership,
+      schoolMembership,
+    };
+  },
+});
+
+/**
+ * Resolve the full route state for one class page.
+ *
+ * The class shell needs one stable backend contract so the frontend can render
+ * either the joined class experience or the join screen without stitching
+ * together multiple client-side queries.
+ */
+export const getClassRoute = query({
+  args: {
+    classId: vv.id("schoolClasses"),
+  },
+  returns: classRouteResultValidator,
+  handler: async (ctx, args) => {
+    const user = await requireAuth(ctx);
+    const classData = await loadActiveClass(ctx, args.classId);
+    const { classMembership, schoolMembership } = await checkClassAccess(
+      ctx,
+      args.classId,
+      classData.schoolId,
+      user.appUser._id
+    );
+
+    if (!schoolMembership) {
+      throw new ConvexError({
+        code: "ACCESS_DENIED",
+        message: "You must be a member of this school to access this class.",
+      });
+    }
+
+    if (classMembership || isAdmin(schoolMembership)) {
+      return {
+        kind: "accessible" as const,
+        class: classData,
+        classMembership,
+        schoolMembership,
+      };
+    }
+
+    return {
+      kind: "joinRequired" as const,
+      class: {
+        _id: classData._id,
+        image: classData.image,
+        name: classData.name,
+        subject: classData.subject,
+        visibility: classData.visibility,
+        year: classData.year,
+      },
       schoolMembership,
     };
   },
